refactor(mocks): extract runs module from mockApi (#991 M3)
Move the run helper listRunsFor and the sixteen run methods (createRun,
createCIFixRun, listRuns, getUsage, getAdminUsage, getMyRateLimits,
getAdminRateLimits, getRun, setRunWaitOnLimit, setRunMrRework, resumeRunNow,
expediteRun, getRunMessages, getRunInputs, submitRunInput, adminListRuns)
verbatim into web/src/mocks/mockApi/runs.ts as runsApi. adminListWorkers stays
in index.ts (a workers method that sat interleaved among the run methods).

Cross-module read per PRD fact 3: getRun builds own_agents from templates
filtered by LEAD_NAME_RE (import from ./agents). runs also pulls
mockOtherRunOwners/mockRunInputs/runListItem etc. from ../data and
ensureLive/handleInput/startNewRun/nextRunId/state from the engine/store.

noUnusedLocals then dropped from index.ts: templates/LEAD_NAME_RE (./agents),
isTerminalRun (../../lib/runStatus, whole import — its leaf-module rationale
comment trimmed to name only the still-imported ApiError), and the
LIVE_RUN_ID/mockAdminRateLimits/mockMyRateLimitsByUser/mockOtherRunOwners/
mockRunInputs/runListItem (../data) and ensureLive/startNewRun (../engine)
bindings plus the AgentSelectionInput/RunPriority/RunInputKind types.
